refactor(Service, util): Service, util 리펙터링
1. LottoMainService 리팩터링
2. 총 수익률 기능을 util로 이동
