+++ src/main/java/lotto/service/LottoPurchaseService.java
@@ -1,0 +1,10 @@
+package lotto.service;
+
+import lotto.model.Lottos;
+
+public class LottoPurchaseService {
+    public static Lottos purchaseLottos(long inputPurchaseAmount) {
+        long lottoCount = inputPurchaseAmount / 1_000;
+        return Lottos.createRandomLottos(lottoCount);
+    }
+}
