+++ src/main/java/lotto/util/RateOfReturnCalculator.java
@@ -1,10 +1,9 @@
-package lotto.service;
+package lotto.util;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 
-public class RateOfReturnService {
-
+public class RateOfReturnCalculator {
     public static double calculateRateOfReturn(long inputLottoPurchaseMoney, long totalLottoWinningAmount) {
         double rateOfReturn = (double) (totalLottoWinningAmount - inputLottoPurchaseMoney) / (inputLottoPurchaseMoney);
         BigDecimal rateOfReturnBigDe = new BigDecimal(rateOfReturn).setScale(1, RoundingMode.HALF_EVEN);
